backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.4
reportlab==4.0.8
Pillow==10.1.0
